Código/requirements.txt
Flask==2.1.2
paho-mqtt==1.6.1
gpxpy==1.5.0
geopy==2.3.0
